# Make whitelistening more strict and explicit