--- a/Gemfile
+++ b/Gemfile
@@ -1,8 +1,8 @@
 source "https://rubygems.org"
 
 group :test, :development do
-  gem 'rspec'
   gem 'pry'
+  gem 'rspec'
 end
 
 gemspec