--- a/lib/logasm/preprocessors/whitelist.rb
+++ b/lib/logasm/preprocessors/whitelist.rb
@@ -1,28 +1,23 @@
+require 'logasm/preprocessors/json_pointer_trie'
+
 class Logasm
   module Preprocessors
     class Whitelist
-
-      DEFAULT_WHITELIST = %w(/id /message /correlation_id /queue)
-      MASK_SYMBOL = '*'
+      DEFAULT_WHITELIST = %w[/id /message /correlation_id /queue].freeze
+      MASK_SYMBOL = '*'.freeze
       MASKED_VALUE = MASK_SYMBOL * 5
-      WILDCARD = '~'
 
       class InvalidPointerFormatException < Exception
       end
 
       def initialize(config = {})
         pointers = (config[:pointers] || []) + DEFAULT_WHITELIST
-        decoded_pointers = pointers
-          .each(&method(:validate_pointer))
-          .map(&method(:decode))
-        @fields_to_include = decoded_pointers.inject({}) do |mem, pointer|
-          mem.merge(pointer => true)
+
+        @trie = pointers.reduce(JSONPointerTrie.new(config)) do |trie, pointer|
+          validate_pointer(pointer)
+
+          trie.insert(decode(pointer))
         end
-        @wildcards = decoded_pointers
-          .select(&method(:has_wildcard?))
-          .inject({}) do |mem, pointer|
-            mem.merge(get_wildcard_roots_of(pointer))
-          end
       end
 
       def process(data)
@@ -31,38 +26,9 @@ class Logasm
 
       private
 
-
-      def has_wildcard?(pointer)
-        pointer.include?("/#{WILDCARD}/") || pointer.end_with?("/#{WILDCARD}")
-      end
-
-      # From a pointer with wildcards builds a hash with roots that contains a wildcard. Hash is used to easily match
-      # find if hash element matches the pointer while processing the log.
-      #
-      # Example:
-      #
-      # Input:
-      #   "/array/~/nested_array/~/fields"
-      #
-      # Output:
-      # {
-      #   "/array/~/nested_array/~" => true,
-      #   "/array/~" => true
-      # }
-      #
-      def get_wildcard_roots_of(pointer)
-        if (index = pointer.rindex("/#{WILDCARD}/"))
-          wildcard_root = pointer.slice(0, index + 2)
-          wildcard_path = pointer.end_with?(WILDCARD) ? pointer : wildcard_root
-          get_wildcard_roots_of(wildcard_root).merge(wildcard_path => true)
-        else
-          {pointer => true}
-        end
-      end
-
       def validate_pointer(pointer)
         if pointer.slice(-1) == '/'
-          raise InvalidPointerFormatException.new('Pointer should not contain trailing slash')
+          raise InvalidPointerFormatException, 'Pointer should not contain trailing slash'
         end
       end
 
@@ -73,52 +39,30 @@ class Logasm
       end
 
       def process_data(parent_pointer, data)
-        self.send("process_#{get_type(data)}", parent_pointer, data)
-      end
+        return MASKED_VALUE unless @trie.include?(parent_pointer)
 
-      def get_type(data)
-        if data.is_a? Hash
-          'hash'
-        elsif data.is_a? Array
-          'array'
+        case data
+        when Hash
+          process_hash(parent_pointer, data)
+
+        when Array
+          process_array(parent_pointer, data)
+
         else
-          'value'
+          data
         end
       end
 
       def process_hash(parent_pointer, hash)
-        create_child_pointer =
-          if @wildcards["#{parent_pointer}/~"]
-            lambda { |_| "#{parent_pointer}/~" }
-          else
-            lambda { |key| "#{parent_pointer}/#{key}" }
-          end
-        hash.inject({}) do |mem, (key, value)|
-          pointer = create_child_pointer.call(key)
-          processed_value = process_data(pointer, value)
-          mem.merge(key => processed_value)
+        hash.each_with_object({}) do |(key, value), result|
+          processed = process_data("#{parent_pointer}/#{key}", value)
+          result[key] = processed
         end
       end
 
       def process_array(parent_pointer, array)
-        create_child_pointer =
-          if @wildcards["#{parent_pointer}/~"]
-            lambda { |_| "#{parent_pointer}/~" }
-          else
-            lambda { |index| "#{parent_pointer}/#{index}" }
-          end
-        array.each_with_index.inject([]) do |mem, (value, index)|
-          pointer = create_child_pointer.call(index)
-          processed_value = process_data(pointer, value)
-          mem + [processed_value]
-        end
-      end
-
-      def process_value(parent_pointer, value)
-        if @fields_to_include[parent_pointer]
-          value
-        else
-          MASKED_VALUE
+        array.each_with_index.map do |value, index|
+          process_data("#{parent_pointer}/#{index}", value)
         end
       end
     end