--- a/spec/preprocessors/whitelist_spec.rb
+++ b/spec/preprocessors/whitelist_spec.rb
@@ -4,24 +4,24 @@ require_relative '../../lib/logasm/preprocessors/whitelist'
 describe Logasm::Preprocessors::Whitelist do
   subject(:processed_data) { described_class.new(config).process(data) }
 
-  let(:config) { {pointers: pointers} }
+  let(:config) { { pointers: pointers } }
   let(:pointers) { [] }
-  let(:data) {{
-    field: 'secret',
-    data: {
-      field: 'secret'
-    },
-    array: [{field: 'secret'}]
-  }}
+  let(:data) do
+    {
+      field: 'secret',
+      data: {
+        field: 'secret'
+      },
+      array: [{ field: 'secret' }]
+    }
+  end
 
   it 'masks all non-whitelisted fields' do
-    expect(processed_data).to eq({
+    expect(processed_data).to eq(
       field: '*****',
-      data: {
-        field: '*****'
-      },
-      array: [{field: '*****'}]
-    })
+      data: '*****',
+      array: '*****'
+    )
   end
 
   context 'when pointer has trailing slash' do
@@ -36,13 +36,11 @@ describe Logasm::Preprocessors::Whitelist do
     let(:pointers) { ['/field'] }
 
     it 'includes the field' do
-      expect(processed_data).to eq({
+      expect(processed_data).to eq(
         field: 'secret',
-        data: {
-          field: '*****'
-        },
-        array: [{field: '*****'}]
-      })
+        data: '*****',
+        array: '*****'
+      )
     end
   end
 
@@ -50,13 +48,13 @@ describe Logasm::Preprocessors::Whitelist do
     let(:pointers) { ['/data/field'] }
 
     it 'includes nested field' do
-      expect(processed_data).to eq({
+      expect(processed_data).to eq(
         field: '*****',
         data: {
           field: 'secret'
         },
-        array: [{field: '*****'}]
-      })
+        array: '*****'
+      )
     end
   end
 
@@ -64,62 +62,70 @@ describe Logasm::Preprocessors::Whitelist do
     let(:pointers) { ['/array/0/field'] }
 
     it 'includes array element' do
-      expect(processed_data).to eq({
+      expect(processed_data).to eq(
         field: '*****',
-        data: {
-          field: '*****'
-        },
-        array: [{field: 'secret'}]
-      })
+        data: '*****',
+        array: [{ field: 'secret' }]
+      )
     end
   end
 
   context 'with whitelisted hash' do
     it 'includes all whitelisted hash elements' do
-      source = {foo: {bar: 'baz'}}
-      target = {foo: {bar: 'baz'}}
+      source = { foo: { bar: 'baz' } }
+      target = { foo: { bar: 'baz' } }
       expect(process(['/foo/~'], source)).to eq(target)
     end
 
     it 'does not include nested elements' do
-      source = {foo: {bar: {baz: 'asd'}}}
-      target = {foo: {bar: {baz: '*****'}}}
+      source = { foo: { bar: { baz: 'asd' } } }
+      target = { foo: { bar: { baz: '*****' } } }
       expect(process(['/foo/~'], source)).to eq(target)
     end
   end
 
   context 'with whitelisted array elements field with wildcard' do
-    let(:data) {{
-      array: [{field: 'data1', secret: 'secret1'}, {field: 'data2', secret: 'secret2'}]
-    }}
+    let(:data) do
+      {
+        array: [
+          { field: 'data1', secret: 'secret1' },
+          { field: 'data2', secret: 'secret2' }
+        ]
+      }
+    end
     let(:pointers) { ['/array/~/field'] }
 
     it 'includes array elements field' do
       expect(processed_data).to include(
-        array: [{field: 'data1', secret: '*****'}, {field: 'data2', secret: '*****'}]
+        array: [
+          { field: 'data1', secret: '*****' },
+          { field: 'data2', secret: '*****' }
+        ]
       )
     end
   end
 
   context 'with whitelisted string array elements with wildcard' do
-    let(:data) {{
-      array: ['secret', 'secret']
-    }}
+    let(:data) do
+      { array: %w[secret secret] }
+    end
     let(:pointers) { ['/array/~'] }
 
     it 'includes array elements' do
-      expect(processed_data).to include(array: ['secret', 'secret'])
+      expect(processed_data).to include(array: %w[secret secret])
     end
   end
 
   context 'with whitelisted string array elements in an array with wildcard' do
-    let(:data) {{
-      nested: [{array: ['secret', 'secret']}]
-    }}
+    let(:data) do
+      {
+        nested: [{ array: %w[secret secret] }]
+      }
+    end
     let(:pointers) { ['/nested/~/array/~'] }
 
     it 'includes array elements' do
-      expect(processed_data).to include(nested: [{array: ['secret', 'secret']}])
+      expect(processed_data).to include(nested: [{ array: %w[secret secret] }])
     end
   end
 
@@ -128,7 +134,7 @@ describe Logasm::Preprocessors::Whitelist do
     let(:pointers) { ['/array/0'] }
 
     it 'masks array element' do
-      expect(processed_data).to include(array: [{field: '*****'}])
+      expect(processed_data).to include(array: [{ field: '*****' }])
     end
   end
 
@@ -136,7 +142,7 @@ describe Logasm::Preprocessors::Whitelist do
     let(:pointers) { ['/array'] }
 
     it 'masks array' do
-      expect(processed_data).to include(array: [{field: '*****'}])
+      expect(processed_data).to include(array: ['*****'])
     end
   end
 
@@ -144,12 +150,12 @@ describe Logasm::Preprocessors::Whitelist do
     let(:pointers) { ['/data'] }
 
     it 'masks hash' do
-      expect(processed_data).to include(data: {field: '*****'})
+      expect(processed_data).to include(data: { field: '*****' })
     end
   end
 
   context 'when boolean present' do
-    let(:data) { {bool: true} }
+    let(:data) { { bool: true } }
 
     it 'masks it with asteriks' do
       expect(processed_data).to eq(bool: '*****')
@@ -157,9 +163,9 @@ describe Logasm::Preprocessors::Whitelist do
   end
 
   context 'when field has slash in the name' do
-    let(:data) {{
-      'field_with_/' => 'secret'
-    }}
+    let(:data) do
+      { 'field_with_/' => 'secret' }
+    end
     let(:pointers) { ['/field_with_~1'] }
 
     it 'includes field' do
@@ -168,31 +174,9 @@ describe Logasm::Preprocessors::Whitelist do
   end
 
   context 'when field has tilde in the name' do
-    let(:data) {{
-      'field_with_~' => 'secret'
-    }}
-    let(:pointers) { ['/field_with_~0'] }
-
-    it 'includes field' do
-      expect(processed_data).to include('field_with_~'=> 'secret')
+    let(:data) do
+      { 'field_with_~' => 'secret' }
     end
-  end
-
-  context 'when field has slash in the name' do
-    let(:data) {{
-      'field_with_/' => 'secret'
-    }}
-    let(:pointers) { ['/field_with_~1'] }
-
-    it 'includes field' do
-      expect(processed_data).to include('field_with_/'=> 'secret')
-    end
-  end
-
-  context 'when field has tilde in the name' do
-    let(:data) {{
-      'field_with_~' => 'secret'
-    }}
     let(:pointers) { ['/field_with_~0'] }
 
     it 'includes field' do
@@ -201,9 +185,9 @@ describe Logasm::Preprocessors::Whitelist do
   end
 
   context 'when field has tilde and 1' do
-    let(:data) {{
-      'field_with_~1' => 'secret'
-    }}
+    let(:data) do
+      { 'field_with_~1' => 'secret' }
+    end
     let(:pointers) { ['/field_with_~01'] }
 
     it 'includes field' do