Previously, whitelistening preprocessor exposed complex fields
structure even they weren't whitelisted. For example, one could
see `{"key": {"foo": "*****", "bar": "*****"}}` in the logs, despite
the fact that `/key` wasn't whitelisted.

This commit makes it so such fields won't be exposed at all, and
example above will produce `{"key":"*****"}`. Also I changed existing
implementation to use Trie data structure, which is much simpler
and (didn't measure) should be faster.

This will require a minor version bump.