@take-five 2017-12-19
@deiwin
Benchmarks.

Old version:
```
Scalar value whitelisting
                         79.890k (± 4.0%) i/s -    401.316k in   5.031461s
Flat hash whitelisting
                         45.661k (± 6.4%) i/s -    230.877k in   5.079152s
Nested hash whitelisting
                         13.231k (± 3.8%) i/s -     66.350k in   5.022311s
Flat array whitelisting
                         20.840k (± 6.9%) i/s -    105.519k in   5.090336s
```

New version
```
Scalar value whitelisting
                        125.807k (± 6.4%) i/s -    626.496k in   5.004610s
Flat hash whitelisting
                         51.643k (± 5.4%) i/s -    259.794k in   5.046759s
Nested hash whitelisting
                         11.224k (± 3.7%) i/s -     56.628k in   5.052578s
Flat array whitelisting
                         14.449k (±17.4%) i/s -     69.414k in   5.068867s
```

Nested hashes became a bit worse, but not significantly. Also, new implementation won't traverse entire structure, so we shouldn't notice any difference