@take-five 2017-12-19
Also, I tried to add cache to Trie implementation and results now are:
```
Calculating -------------------------------------
Scalar value whitelisting
                        234.343k (± 4.8%) i/s -      1.171M in   5.010703s
Flat hash whitelisting
                        108.316k (± 4.1%) i/s -    550.165k in   5.087940s
Nested hash whitelisting
                         25.407k (± 3.4%) i/s -    127.452k in   5.022319s
Flat array whitelisting
                         35.661k (± 4.4%) i/s -    180.284k in   5.066247s
```